+++ ofrak_core/tests/components/test_comments.py
@@ -1,3 +1,7 @@
+"""
+Test the functionality of comments in OFRAK Core, including adding, modifying
+and removing comments on resources.
+"""
 from datetime import timedelta
 import pytest
 from hypothesis import HealthCheck, given, settings