+++ ofrak_core/tests/components/test_linkable_symbol.py
@@ -1,3 +1,9 @@
+"""
+Test the functionality of linkable symbols.
+
+Requirements Mapping:
+- REQ1.2
+"""
 import pytest
 
 from ofrak.core import LinkableSymbol, LinkableSymbolType, LinkableSymbolStubInfo
@@ -14,6 +20,9 @@ from ofrak.core import LinkableSymbol, LinkableSymbolType, LinkableSymbolStubInf
 def test_linkable_symbol_stub_info(linkable_symbol: LinkableSymbol):
     """
     Test that it is possible to generate LinkableSymbolStubInfo for given linkable sybmol.
+
+    This test verifies that:
+    - A LinkableSymbolStubInfo can be generated for various symbol types
     """
     stub_info = linkable_symbol.get_stub_info()
     assert isinstance(stub_info, LinkableSymbolStubInfo)
@@ -22,6 +31,9 @@ def test_linkable_symbol_stub_info(linkable_symbol: LinkableSymbol):
 def test_linkable_symbol_stub_info_undef():
     """
     Test that get_stub_info raises NotImplementedError for LinkableSymbolType.UNDEF.
+
+    This test verifies that:
+    - Attempting to generate stub info for an undefined symbol type raises NotImplementedError
     """
     linkable_symbol = LinkableSymbol(0x1337, "leet", LinkableSymbolType.UNDEF)
     with pytest.raises(NotImplementedError):