+++ ofrak_core/tests/components/test_xattr_stub.py
@@ -1,3 +1,9 @@
+"""
+This module tests the functionality of the xattr stub implementation.
+
+Requirements Mapping:
+- REQ1.2
+"""
 import os
 
 import pytest
@@ -15,60 +21,90 @@ def xattr_class_stub():
 
 
 def test_repr(xattr_class_stub, caplog):
+    """
+    This test verifies that the __repr__ method returns a string representation of the xattr object.
+    """
     value = xattr_class_stub.__repr__()
     assert "Function __repr__" in caplog.text
     assert isinstance(value, str)
 
 
 def test_get(xattr_class_stub, caplog):
+    """
+    This test verifies that the get method retrieves extended attributes.
+    """
     value = xattr_class_stub.get("user.foo")
     assert "Function get" in caplog.text
     assert value == b""
 
 
 def test_set(xattr_class_stub, caplog):
+    """
+    This test verifies that the set method sets extended attributes.
+    """
     value = xattr_class_stub.set("user.baz", b"qux")
     assert "Function set" in caplog.text
     assert value is None
 
 
 def test_remove(xattr_class_stub, caplog):
+    """
+    This test verifies that the remove method removes extended attributes.
+    """
     value = xattr_class_stub.remove("user.foo")
     assert "Function remove" in caplog.text
     assert value is None
 
 
 def test_list(xattr_class_stub, caplog):
+    """
+    This test verifies that the list method lists extended attributes.
+    """
     value = xattr_class_stub.list()
     assert "Function list" in caplog.text
     assert len(value) == 0
 
 
 def test_len(xattr_class_stub, caplog):
+    """
+    This test verifies that the __len__ method returns the number of extended attributes.
+    """
     value = xattr_class_stub.__len__()
     assert "Function __len__" in caplog.text
     assert value == 0
 
 
 def test_delitem(xattr_class_stub, caplog):
+    """
+    This test verifies that the __delitem__ method removes an extended attribute.
+    """
     value = xattr_class_stub.__delitem__("user.foo")
     assert "Function __delitem__" in caplog.text
     assert value is None
 
 
 def test_setitem(xattr_class_stub, caplog):
+    """
+    This test verifies that the __setitem__ method sets an extended attribute.
+    """
     value = xattr_class_stub.__setitem__("user.baz", b"qux")
     assert "Function __setitem__" in caplog.text
     assert value is None
 
 
 def test_getitem(xattr_class_stub, caplog):
+    """
+    This test verifies that the __getitem__ method retrieves an extended attribute.
+    """
     value = xattr_class_stub.__getitem__("user.foo")
     assert "Function __getitem__" in caplog.text
     assert value == b""
 
 
 def test_iterkeys(xattr_class_stub, caplog):
+    """
+    This test verifies that the iterkeys method iterates over extended attribute keys.
+    """
     value = xattr_class_stub.iterkeys()
     assert "Function iterkeys" in caplog.text
     with pytest.raises(StopIteration):
@@ -76,84 +112,126 @@ def test_iterkeys(xattr_class_stub, caplog):
 
 
 def test_has_key(xattr_class_stub, caplog):
+    """
+    This test verifies that the has_key method checks for the existence of an extended attribute.
+    """
     value = xattr_class_stub.has_key("user.foo")
     assert "Function has_key" in caplog.text
     assert value is False
 
 
 def test_clear(xattr_class_stub, caplog):
+    """
+    This test verifies that the clear method removes all extended attributes.
+    """
     value = xattr_class_stub.clear()
     assert "Function clear" in caplog.text
     assert value is None
 
 
 def test_update(xattr_class_stub, caplog):
+    """
+    This test verifies that the update method updates extended attributes.
+    """
     value = xattr_class_stub.update({"user.foo": b"baz"})
     assert "Function update" in caplog.text
     assert value is None
 
 
 def test_copy(xattr_class_stub, caplog):
+    """
+    This test verifies that the copy method creates a copy of the xattr object.
+    """
     value = xattr_class_stub.copy()
     assert "Function copy" in caplog.text
     assert len(value) == 0
 
 
 def test_setdefault(xattr_class_stub, caplog):
+    """
+    This test verifies that the setdefault method sets a default extended attribute.
+    """
     value = xattr_class_stub.setdefault("user.foo")
     assert "Function setdefault" in caplog.text
     assert value == b""
 
 
 def test_keys(xattr_class_stub, caplog):
+    """
+    This test verifies that the keys method returns a list of extended attribute keys.
+    """
     value = xattr_class_stub.keys()
     assert "Function keys" in caplog.text
     assert len(value) == 0
 
 
 def test_itervalues(xattr_class_stub, caplog):
+    """
+    This test verifies that the itervalues method iterates over extended attribute values.
+    """
     value = xattr_class_stub.itervalues()
     assert next(value) == b""
     assert "Function itervalues" in caplog.text
 
 
 def test_values(xattr_class_stub, caplog):
+    """
+    This test verifies that the values method returns a list of extended attribute values.
+    """
     value = xattr_class_stub.values()
     assert "Function values" in caplog.text
     assert len(value) == 0
 
 
 def test_iteritems(xattr_class_stub, caplog):
+    """
+    This test verifies that the iteritems method iterates over extended attribute key-value pairs.
+    """
     value = xattr_class_stub.iteritems()
     assert next(value) == ()
     assert "Function iteritems" in caplog.text
 
 
 def test_items(xattr_class_stub, caplog):
+    """
+    This test verifies that the items method returns a list of extended attribute key-value pairs.
+    """
     value = xattr_class_stub.items()
     assert "Function items" in caplog.text
     assert len(value) == 0
 
 
 def test_listxattr(caplog):
+    """
+    This test verifies that the listxattr function lists extended attributes on a file.
+    """
     value = xattr_stub.listxattr(EXAMPLE_FILE)
     assert "Function listxattr" in caplog.text
     assert value == ()
 
 
 def test_getxattr(caplog):
+    """
+    This test verifies that the getxattr function retrieves an extended attribute from a file.
+    """
     value = xattr_stub.getxattr(EXAMPLE_FILE, "user.foo")
     assert "Function getxattr" in caplog.text
     assert value == b""
 
 
 def test_setxattr(caplog):
+    """
+    This test verifies that the setxattr function sets an extended attribute on a file.
+    """
     value = xattr_stub.setxattr(EXAMPLE_FILE, "user.baz", b"qux")
     assert "Function setxattr" in caplog.text
     assert value is None
 
 
 def test_removexattr(caplog):
+    """
+    This test verifies that the removexattr function removes an extended attribute from a file.
+    """
     value = xattr_stub.removexattr(EXAMPLE_FILE, "user.foo")
     assert "Function removexattr" in caplog.text
     assert value is None