+++ ofrak_patch_maker/tests/test_ppc_toolchain.py
@@ -1,3 +1,6 @@
+"""
+Test the PowerPC toolchain functionality.
+"""
 import pytest
 
 from .toolchain_under_test import ToolchainUnderTest